requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
aiosqlite>=0.19.0

# Auth (JWT + password hashing)
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
